Make SparseArrays.jl a weak dependency
