+++ src/misc.jl
@@ -138,6 +138,7 @@ function indicatormat(x::AbstractArray{<:Integer}, k::Integer; sparse::Bool=fals
     sparse ? _indicatormat_sparse(x, k) : _indicatormat_dense(x, k)
 end
 
+function _indicatormat_sparse end
 
 """
     indicatormat(x, c=sort(unique(x)); sparse=false)
@@ -175,18 +176,4 @@ function _indicatormat_dense(x::AbstractArray{T}, c::AbstractArray{T}) where T
         o += m
     end
     return r
-end
-
-_indicatormat_sparse(x::AbstractArray{<:Integer}, k::Integer) = (n = length(x); sparse(x, 1:n, true, k, n))
-
-function _indicatormat_sparse(x::AbstractArray{T}, c::AbstractArray{T}) where T
-    d = indexmap(c)
-    m = length(c)
-    n = length(x)
-
-    rinds = Vector{Int}(undef, n)
-    @inbounds for i = 1 : n
-        rinds[i] = d[x[i]]
-    end
-    return sparse(rinds, 1:n, true, m, n)
 end
